usersession: move autostart code to usersession/autostart
